rag-multilang-strands/requirements.txt
strands-agents
streamlit
boto3
langdetect
faiss-cpu
pypdf2
python-docx
